requirements.txt
# FastAPI and server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# HTTP client
httpx>=0.26.0

# Data processing
pandas>=2.2.0
numpy>=1.26.0

# Machine Learning
scikit-learn>=1.4.0

# Environment variables
python-dotenv>=1.0.0

# Pydantic (included with FastAPI, but explicit for schemas)
pydantic>=2.5.0
